chore: create one component to handle credit notes list
