+++ ditto/base.json
@@ -238,14 +238,8 @@
   "text_63725b30957fd5b26b308dd9": "Total amount available",
   "text_63725b30957fd5b26b308ddb": "Sum of {{count}} credit notes credited",
   "text_63725b30957fd5b26b308ddf": "Credit notes",
-  "text_63725b30957fd5b26b308de1": "Issuing date",
-  "text_63725b30957fd5b26b308de3": "Credit note number",
-  "text_63725b30957fd5b26b308de5": "Amount",
   "text_63728c6434e1344aea76347d": "Download, void, copy ID",
   "text_63728c6434e1344aea76347f": "Download, copy ID",
-  "text_63725b30957fd5b26b308df9": "Download pdf",
-  "text_63725b30957fd5b26b308dfb": "Void credit available",
-  "text_63725b30957fd5b26b308dfd": "Copy credit note lago_id",
   "text_634687079be251fdb438338f": "Actions",
   "text_634687079be251fdb4383395": "Download invoice",
   "text_634687079be251fdb438339b": "Copy invoice_id",
